requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
elasticsearch
neo4j
numpy
pandas
scikit-learn
